# design/openapi.yaml
...
openapi: 3.0.3
info:
  title: DriveDeal — Reverse Car Marketplace API
  version: 1.0.0
  description: |-
    Backend API for DriveDeal: a reverse car marketplace where buyers post the
    car they want and dealers compete on price for it.

    Handles:
    - Buyer flows: post a car request, watch the live quote leaderboard, accept a
      quote, track the deal, chat with the dealer.
    - Dealer flows: account with manual approval, geo-filtered request feed, quote
      submission and revision, deal status.
    - Support flows: ticket queue, dealer/agent verification approval, roster.
    - AI flows (buyer only): a streaming advisory agent and a compare agent,
      grounded in the database plus a Crawl4AI web search.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the /api/v1/ws WebSocket.
    - **Success responses are envelope-free.** The payload is the resource itself.
      There is no {"success": true, "data": ...} wrapper — the frontend's mock
      client already returns bare objects and that shape is the contract
      (solution-frontend.md §2.1, solution-backend.md §0.2).
    - **Every non-2xx JSON response uses the error envelope**, and nothing else:
        {"error": {"code", "message", "details", "request_id"}}
      `code` is a stable machine-readable string the client branches on.
      `message` is for humans and may be reworded; never parse it.
...
  contact:
...
servers:
...
security:
- bearer_auth: []
paths:
...
  /requests:
    post:
      summary: Create a car request
      description: |-
        Buyers describe the car they want; dealers bid on it. Creates a
        `buyer_requests` row in `open` status and copies the buyer's
        `buyer_preference` snapshot onto the row, so later preference edits do
        not silently change what dealers are being asked for.
      operationId: createRequest
      tags:
      - Requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpsert'
            examples:
              seven_seater:
                value:
                  title: 7-seater SUV under 65k
                  description: Looking for a reliable family SUV, third row preferred.
                  state_id: 0e1f2a3b-4c5d-4e6f-8a9b-0c1d2e3f4a5b
                  brand_id: b5a5e627-7f0f-4d6f-a7e5-1b2c3d4e5f60
                  body_type: SUV
                  seater_count: 7
                  min_year: 2021
                  max_mileage: 60000
                  budget_min: '40000.00'
                  budget_max: '65000.00'
                  target_otd: '68000.00'
                  transmission: Automatic
                  drivetrain: AWD
                  fuel_type: Petrol
                  condition: Used
                  exterior_color: Blue
                  must_haves:
                  - sunroof
                  - leather
                  - apple_carplay
                  nice_to_haves:
                  - third_row
                  - heated_seats
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '422':
          $ref: '#/components/responses/validation_error'
    get:
      summary: List my requests
      description: |-
        Cursor-paginated, newest first. Optional `status` filter. `next_cursor`
        is null on the last page.
      operationId: listMyRequests
      tags:
      - Requests
      parameters:
      - name: status
        in: query
        description: Filter by request status
        schema:
          $ref: '#/components/schemas/RequestStatus'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestList'
              examples:
                one_open:
                  value:
                    items:
                    - $ref: '#/components/examples/RequestOpen'
                    next_cursor: null
        '401':
          $ref: '#/components/responses/unauthorized'
  /requests/{id}:
    get:
      summary: Get one of my requests
      description: |-
        Buyer-only: 403 for anyone but the owning buyer. Includes the quote
        count so the list row and the detail header agree.
      operationId: getRequest
      tags:
      - Requests
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerRequestDetail'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
    patch:
      summary: Update one of my requests
      description: |-
        Rejected with 422 once `status` is not `open` — once dealers are
        quoting, the spec is frozen. Close the request to stop it.
      operationId: updateRequest
      tags:
      - Requests
      parameters:
      - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpsert'
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '422':
          $ref: '#/components/responses/validation_error'
  /requests/{id}/publish:
    post:
      summary: Publish an AI-drafted request
      description: |-
        **The human-in-the-loop gate.** Moves `draft → open`, which is the only
        way a request becomes visible in any dealer feed.

        No LangGraph node can call this. The AI's write surface to
        buyer_requests is `draft` only; a test asserts no graph path can reach
        `status = 'open'` (solution-backend.md §10.6).
      operationId: publishRequest
      tags:
      - Requests
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request is now open
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /requests/{id}/close:
    post:
      summary: Close one of my requests
      description: |-
        Buyer-only, terminal transition. `open → closed`, or `open → matched` for
        the buyer who already accepted a quote. Withdrawing every outstanding
        quote is a side effect of matching, done in the same transaction.
      operationId: closeRequest
      tags:
      - Requests
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
...
  /requests/{id}/quotes:
    get:
      summary: Live quote leaderboard for a request
      description: |-
        Every quote on the request, ordered by the composite score the frontend
        renders: `effective_price` ascending, tie-broken by
        `buyer_chat_cleared_at` (dealer's messages visible) then
        `seller_rating` descending.

        Only the owning buyer may read this — it exposes dealer pricing.
      operationId: listRequestQuotes
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IncludeWithdrawn'
      responses:
        '200':
          description: Quotes ordered best-first
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuoteSummary'
              examples:
                two_quotes:
                  value:
                    items:
                    - $ref: '#/components/examples/QuoteLeading'
                    - $ref: '#/components/examples/QuoteSecond'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
  /requests/{id}/compare-ids:
    get:
      summary: List request ids eligible for compare
      description: |-
        The compare agent's picker. Returns the buyer's own requests that are
        `open` and not already in the comparison, so the UI can offer valid
        choices without the client guessing at eligibility rules.
      operationId: listComparableRequests
      tags:
      - AI
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Comparable request ids
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComparableRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
  /feed/requests:
    get:
      summary: Geo-filtered request feed for dealers
      description: |-
        The dealer's core feed. `radius_miles` defaults to 50; `state_id` is an
        additional filter. Matching mirrors the buyer's preference semantics
        exactly — a NULL preference column is a wildcard, so a request the
        dealer satisfies on nothing but state still appears.

        Dealer-only. Support agents get 403 here; their queue is
        /support/queue/tickets.
      operationId: getDealerRequestFeed
      tags:
      - Dealer Feed
      parameters:
      - name: state_id
        in: query
        description: Restrict to one state
        schema:
          type: string
          format: uuid
      - name: radius_miles
        in: query
        description: Search radius around the dealer's registered address
        schema:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
      - name: brand_id
        in: query
        description: Filter to requests wanting this brand
        schema:
          type: string
          format: uuid
      - name: body_type
        in: query
        schema:
          $ref: '#/components/schemas/BodyType'
      - name: budget_max
        in: query
        description: Requests whose budget_max is at or below this
        schema:
          $ref: '#/components/schemas/DecimalString'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of matching requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedRequestList'
              examples:
                page:
                  value:
                    items:
                    - $ref: '#/components/examples/FeedRequest'
                    next_cursor: eyJpZCI6IjAyZTkyYzZhIn0
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /feed/requests/{id}:
    get:
      summary: One request from the dealer feed
      description: |-
        The dealer-facing read of a request. Same row as the buyer's view, but
        the buyer snapshot is trimmed to the fields a dealer needs to price it
        — no contact details, no other dealers' quotes, no internal IDs.
      operationId: getFeedRequest
      tags:
      - Dealer Feed
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request as a dealer sees it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedRequest'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
  /quotes:
    post:
      summary: Quote on a request
      description: |-
        Creates a `deal_quotes` row. Dealer-only, and the dealer must have an
        approved `support_verifications` row — 403 DEALER_PENDING_REVIEW
        otherwise.

        One live quote per (request, dealer) enforced by a partial unique index;
        a second quote for the same pair is 409, not a silent update.

        **final_price is server-computed** as `(price - discount) * (1 +
        sales_tax_rate)`, using the request's state tax rate. Sending
        final_price is 422.
      operationId: createQuote
      tags:
      - Quotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteCreateRequest'
            examples:
              first_offer:
                value:
                  request_id: 2fa7ac94-1b3c-4d5e-8f90-1a2b3c4d5e6f
                  price: '62000.00'
                  discount: '2000.00'
                  estimated_delivery_days: 7
                  message: We have a matching 2022 Bronco in stock, fully inspected.
                  car_ids:
                  - 00e22c60-88c0-40ef-a12d-365ecbbd4185
                  valid_until: '2026-10-07 00:00:00+00:00'
      responses:
        '201':
          description: Quote created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteDetail'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
    get:
      summary: List my quotes
      description: Cursor-paginated, newest first. Dealer-only.
      operationId: listMyQuotes
      tags:
      - Quotes
      parameters:
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/QuoteStatus'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of quotes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
  /quotes/{id}:
    get:
      summary: Get one of my quotes
      description: |-
        Dealer-only. The buyer reads the same quote through
        /requests/{request_id}/quotes, which is a trimmed projection.
      operationId: getQuote
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Quote detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteDetail'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
  /quotes/{id}/revise:
    post:
      summary: Revise my quote
      description: |-
        Replaces price / discount / delivery estimate on a `pending` quote and
        bumps `revision_count`. The prior values are preserved in
        `deal_quotes.revisions` (jsonb array) — a revision is an append, not an
        overwrite, so a dealer cannot erase a number they already showed a
        buyer.
      operationId: reviseQuote
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteReviseRequest'
            examples:
              improved:
                value:
                  price: '59500.00'
                  discount: '2500.00'
                  estimated_delivery_days: 5
                  message: Improved offer — includes a fresh inspection and 12-month warranty.
      responses:
        '200':
          description: Revised quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteDetail'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
  /quotes/{id}/accept:
    post:
      summary: Accept my quote
      description: |-
        Buyer-only, and only on a `pending` quote whose request is still `open`.
        The accept is a compare-and-set: the losing quotes are marked
        `withdrawn` and the request moves to `matched` in the **same
        transaction**, so two buyers can never accept two quotes on one request.

        Idempotent on the same quote — a retry returns 200 with the same
        `deal_id`, not a 409, because a flaky network must not double-create a
        deal.
      operationId: acceptQuote
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Quote accepted, deal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteAcceptResult'
              examples:
                accepted:
                  value:
                    quote_id: 2fa7ac94-1b3c-4d5e-8f90-1a2b3c4d5e6f
                    deal_id: dce16b4b-7c8d-4e9f-a0b1-c2d3e4f5a6b
                    request_status: matched
                    effective_price: '61485.40'
                    withdrawn_quote_count: 2
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /quotes/{id}/dealer-contact:
    get:
      summary: Reveal the dealer contact for this quote
      description: |-
        Buyer-only. This is the endpoint that puts the dealer's phone, email,
        and business name on the buyer's offer page, and it is **gated** — the
        contact stays hidden until one of these two conditions holds:

        1. The buyer accepted the quote (`status = accepted` or `matched`).
        2. The buyer asked to negotiate, the dealer accepted that chat request,
           and the quote is `negotiating`.

        Before either condition, this is 403 `DEALER_CONTACT_WITHHELD` — a
        distinct code so the client can render "accept the offer or request
        contact to see dealer details" instead of a generic failure.

        The chat icon on the buyer's offer page is driven by
        `contact_available` in this response, so the two never disagree: no
        contact means no chat icon.
      operationId: getQuoteDealerContact
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Dealer contact unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerContact'
              examples:
                unlocked:
                  value:
                    quote_id: 2fa7ac94-1b3c-4d5e-8f90-1a2b3c4d5e6f
                    contact_available: true
                    unlocked_by: accepted_quote
                    dealer_profile_id: bdc31c9e-f6ed-40d0-a32f-706062f81d77
                    display_name: Ridgeway Motors
                    avatar_url: https://cdn.drivedeal.example/avatars/bdc31c9e.jpg
                    business_name: Ridgeway Motors LLC
                    phone: '+15550142288'
                    email: deals@ridgewaymotors.example
                    is_verified_dealer: true
                    average_response_minutes: 18
                    chat_url: /api/v1/ws?ticket=<one-shot ticket from POST /chats/ws-ticket>
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          description: Contact still withheld
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                withheld:
                  value:
                    error:
                      code: DEALER_CONTACT_WITHHELD
                      message: Accept the offer or request a chat to see dealer contact details.
                      details:
                        quote_status: pending
                        chat_request_status: null
        '404':
          $ref: '#/components/responses/not_found'
  /quotes/{id}/decline:
    post:
      summary: Decline my quote
      description: Buyer-only. Moves the quote to `declined`; the request stays open.
      operationId: declineQuote
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Quote declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSummary'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /quotes/{id}/withdraw:
    post:
      summary: Withdraw my quote
      description: |-
        Dealer-only, `pending` → `withdrawn`. Allowed even after the buyer has
        opened a chat — the chat row is soft-cleared on the dealer side, not
        deleted (solution-backend.md §9.4).
      operationId: withdrawQuote
      tags:
      - Quotes
      parameters:
      - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Quote withdrawn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSummary'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not_found'
        '409':
          $ref: '#/components/responses/conflict'
  /deals:
    get:
      summary: List my deals
      description: |-
        Deals where the caller is the buyer or the winning dealer. Cursor
        pagination, newest activity first.
      operationId: listMyDeals
      tags:
      - Deals
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealList'
        '401':
          $ref: '#/components/responses/unauthorized'
...
components:
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        RS256 access token from /auth/login or /auth/refresh.

        Subject = `profiles.id` (NOT the `users.id`, which is a different
        UUID and does not appear anywhere in the request layer). Roles are
        read from the `profiles` row on every request, so a role change or a
        suspension takes effect on the next call rather than at token expiry.

        For the WebSocket, use a one-shot ticket from /chats/ws-ticket
        instead — a browser cannot set this header on the handshake.
...